backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.3
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
dlib @ file:///E:/FYP/FYP_AiDriven_Meeting_Analysis/backend/dlib-19.24.1-cp310-cp310-win_amd64.whl
emoji==2.14.1
exceptiongroup==1.3.0
face-recognition @ file:///E:/FYP/FYP_AiDriven_Meeting_Analysis/backend/face_recognition-1.3.0-py2.py3-none-any.whl
face-recognition-models==0.3.0
facenet-pytorch==2.6.0
fastapi==0.115.12
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
future==1.0.0
h11==0.16.0
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.3
joblib==1.5.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
onnxruntime==1.22.0
opencv-python==4.11.0.86
packaging==25.0
pillow==10.2.0
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
pyreadline3==3.5.4
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.16
tokenizers==0.21.1
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
